BOOKMARK/requirements.txt
asgiref==3.2.10
backcall==0.2.0
certifi==2020.11.8
chardet==3.0.4
colorama==0.4.4
decorator==4.4.2
Django==3.1.2
django-crispy-forms==1.9.2
django-debug-toolbar==3.2
django-extensions==3.0.9
django-js-asset==1.2.2
django-mptt==0.11.0
django-restframework==0.0.1
djangorestframework==3.12.2
httpie==2.3.0
idna==2.10
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.17.2
parso==0.7.1
pickleshare==0.7.5
prompt-toolkit==3.0.8
Pygments==2.7.2
PySocks==1.7.1
pytz==2020.1
requests==2.25.0
requests-toolbelt==0.9.1
six==1.15.0
sqlparse==0.4.1
traitlets==5.0.5
urllib3==1.26.2
wcwidth==0.2.5
